NESDL running on Windows (w/o menu bar)! Platform changes, cast ambiguities, CPU perf improvement
  * Updated VCXPROJ to add new SDL_TTF library, new includes and source file subdirectories
    * Updated project  to C++20
    * Added _CRT_SECURE_NO_WARNINGS to remove warnings for strncpy and strncat (used in nativefiledialog)
  * Headers and includes updated for Windows (it's a special case with this one)
  * Attempted implementing a more accurate FPS cap on Windows (more stable but less perf, kept around for future usage)
  * Renamed AddrMode enums due to collision with Windows
  * Added explicit casts for values to clear up compiler ambiguities
  * Cut down on CPU usage by limiting calls to GetCyclesForNextInstruction() (called too often when we could cache the result)
    * Ended up making a difference when Debug profiling on Windows

+++ Source/src/NESDL_RAM.cpp
@@ -32,7 +32,7 @@ uint8_t NESDL_RAM::ReadByte(uint16_t addr)
         bool isPlayer2 = addr == 0x4017;
         // High 3 (or 5?) bytes are open bus, usually containing the upper byte of the controller address
         // (aka 0x40). Needed for Paperboy as it expects this open bus data, according to NESDEV
-        return 0x40 | core->input->GetNextPlayerInputBit(isPlayer2);
+        return 0x40 | (uint8_t)core->input->GetNextPlayerInputBit(isPlayer2);
     }
     // APU test functionality + some unfinished hardware functionality (do nothing)
     else if (addr < 0x4020)
